chore(skore): reorganize tests folders to reflect the skore API
