+++ skore/tests/unit/displays/conftest.py
