+++ skore/tests/unit/displays/confusion_matrix/test_estimator.py
