+++ skore/tests/unit/displays/precision_recall_curve/test_common.py
