+++ skore/tests/unit/displays/precision_recall_curve/test_comparison_cross_validation.py
