+++ skore/tests/unit/displays/precision_recall_curve/test_comparison_estimator.py
