+++ skore/tests/unit/displays/precision_recall_curve/test_cross_validation.py
