+++ skore/tests/unit/displays/prediction_error/conftest.py
