+++ skore/tests/unit/displays/prediction_error/test_comparison_estimator.py
