+++ skore/tests/unit/displays/prediction_error/test_cross_validation.py
