+++ skore/tests/unit/displays/prediction_error/test_estimator.py
