+++ skore/tests/unit/displays/roc_curve/test_common.py
