+++ skore/tests/unit/displays/roc_curve/test_comparison_cross_validation.py
