+++ skore/tests/unit/displays/roc_curve/test_comparison_estimator.py
