+++ skore/tests/unit/displays/roc_curve/test_cross_validation.py
