+++ skore/tests/unit/displays/roc_curve/test_estimator.py
