+++ skore/tests/unit/displays/table_report/test_estimator.py
