+++ skore/tests/unit/displays/test_common.py
