+++ skore/tests/unit/displays/test_style.py
