+++ skore/tests/unit/displays/test_utils.py
