+++ skore/tests/unit/reports/estimator/feature_importance/test_accessor.py
@@ -1,11 +1,10 @@
-from sklearn.linear_model import LinearRegression
 from skore import EstimatorReport
 
 
-def test_estimator_report_feature_importance_help(capsys, regression_data):
+def test_estimator_report_feature_importance_help(capsys, linear_regression_with_test):
     """Check that the help method writes to the console."""
-    X, y = regression_data
-    report = EstimatorReport(LinearRegression().fit(X, y), X_test=X, y_test=y)
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     report.feature_importance.help()
     captured = capsys.readouterr()
@@ -13,10 +12,10 @@ def test_estimator_report_feature_importance_help(capsys, regression_data):
     assert "coefficients" in captured.out
 
 
-def test_estimator_report_feature_importance_repr(regression_data):
+def test_estimator_report_feature_importance_repr(linear_regression_with_test):
     """Check that __repr__ returns a string starting with the expected prefix."""
-    X, y = regression_data
-    report = EstimatorReport(LinearRegression().fit(X, y), X_test=X, y_test=y)
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     repr_str = repr(report.feature_importance)
     assert "skore.EstimatorReport.feature_importance" in repr_str
