+++ skore/tests/unit/reports/estimator/metrics/test_numeric.py
@@ -1,18 +1,14 @@
-import inspect
 import re
-from copy import deepcopy
-from io import BytesIO
 from numbers import Real
 
 import joblib
 import numpy as np
 import pandas as pd
 import pytest
-from sklearn.base import clone
+from sklearn.base import BaseEstimator, clone
 from sklearn.cluster import KMeans
-from sklearn.datasets import make_classification, make_regression
-from sklearn.ensemble import RandomForestClassifier
-from sklearn.linear_model import LinearRegression, LogisticRegression
+from sklearn.datasets import make_classification
+from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import (
     accuracy_score,
     f1_score,
@@ -25,578 +21,13 @@ from sklearn.metrics import (
     recall_score,
 )
 from sklearn.model_selection import train_test_split
-from sklearn.pipeline import Pipeline
-from sklearn.preprocessing import StandardScaler
 from sklearn.svm import SVC
-from sklearn.utils.validation import check_is_fitted
 from skore import EstimatorReport
-from skore._sklearn._plot import RocCurveDisplay
 
 
-@pytest.fixture
-def binary_classification_data():
-    """Create a binary classification dataset and return fitted estimator and data."""
-    X, y = make_classification(random_state=42)
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, test_size=0.2, random_state=42
-    )
-    return RandomForestClassifier().fit(X_train, y_train), X_test, y_test
-
-
-@pytest.fixture
-def binary_classification_data_svc():
-    """Create a binary classification dataset and return fitted estimator and data.
-    The estimator is a SVC that does not support `predict_proba`.
-    """
-    X, y = make_classification(random_state=42)
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, test_size=0.2, random_state=42
-    )
-    return SVC().fit(X_train, y_train), X_test, y_test
-
-
-@pytest.fixture
-def multiclass_classification_data():
-    """Create a multiclass classification dataset and return fitted estimator and
-    data."""
-    X, y = make_classification(
-        n_classes=3, n_clusters_per_class=1, random_state=42, n_informative=10
-    )
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, test_size=0.2, random_state=42
-    )
-    return RandomForestClassifier().fit(X_train, y_train), X_test, y_test
-
-
-@pytest.fixture
-def multiclass_classification_data_svc():
-    """Create a multiclass classification dataset and return fitted estimator and
-    data. The estimator is a SVC that does not support `predict_proba`.
-    """
-    X, y = make_classification(
-        n_classes=3, n_clusters_per_class=1, random_state=42, n_informative=10
-    )
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, test_size=0.2, random_state=42
-    )
-    return SVC().fit(X_train, y_train), X_test, y_test
-
-
-@pytest.fixture
-def binary_classification_data_pipeline():
-    """Create a binary classification dataset and return fitted pipeline and data."""
-    X, y = make_classification(random_state=42)
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, test_size=0.2, random_state=42
-    )
-    estimator = Pipeline([("scaler", StandardScaler()), ("clf", LogisticRegression())])
-    return estimator.fit(X_train, y_train), X_test, y_test
-
-
-@pytest.fixture
-def regression_data():
-    """Create a regression dataset and return fitted estimator and data."""
-    X, y = make_regression(random_state=42)
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, test_size=0.2, random_state=42
-    )
-    return LinearRegression().fit(X_train, y_train), X_test, y_test
-
-
-@pytest.fixture
-def regression_multioutput_data():
-    """Create a regression dataset and return fitted estimator and data."""
-    X, y = make_regression(n_targets=2, random_state=42)
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, test_size=0.2, random_state=42
-    )
-    return LinearRegression().fit(X_train, y_train), X_test, y_test
-
-
-########################################################################################
-# Check the general behaviour of the report
-########################################################################################
-
-
-def test_report_can_be_rebuilt_using_parameters(regression_data):
-    estimator, X_test, y_test = regression_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-    parameters = {}
-
-    for parameter in inspect.signature(EstimatorReport).parameters:
-        assert hasattr(report, parameter), f"The parameter '{parameter}' must be stored"
-
-        parameters[parameter] = getattr(report, parameter)
-
-    EstimatorReport(**parameters)
-
-
-@pytest.mark.parametrize("fit", [True, "auto"])
-def test_estimator_not_fitted(fit):
-    """Test that an error is raised when trying to create a report from an unfitted
-    estimator and no data are provided to fit the estimator.
-    """
-    estimator = LinearRegression()
-    err_msg = "The training data is required to fit the estimator. "
-    with pytest.raises(ValueError, match=err_msg):
-        EstimatorReport(estimator, fit=fit)
-
-
-@pytest.mark.parametrize("fit", [True, "auto"])
-def test_estimator_report_from_unfitted_estimator(fit):
-    """Check the general behaviour of passing an unfitted estimator and training
-    data."""
-    X, y = make_regression(random_state=42)
-    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)
-    estimator = LinearRegression()
-    report = EstimatorReport(
-        estimator,
-        fit=fit,
-        X_train=X_train,
-        y_train=y_train,
-        X_test=X_test,
-        y_test=y_test,
-    )
-
-    check_is_fitted(report.estimator_)
-    assert report.estimator_ is not estimator  # the estimator should be cloned
-
-    assert report.X_train is X_train
-    assert report.y_train is y_train
-    assert report.X_test is X_test
-    assert report.y_test is y_test
-
-    with pytest.raises(AttributeError):
-        report.estimator_ = LinearRegression()
-    with pytest.raises(AttributeError):
-        report.X_train = X_train
-    with pytest.raises(AttributeError):
-        report.y_train = y_train
-
-
-@pytest.mark.parametrize("fit", [False, "auto"])
-def test_estimator_report_from_fitted_estimator(binary_classification_data, fit):
-    """Check the general behaviour of passing an already fitted estimator without
-    refitting it."""
-    estimator, X, y = binary_classification_data
-    report = EstimatorReport(estimator, fit=fit, X_test=X, y_test=y)
-
-    check_is_fitted(report.estimator_)
-    assert isinstance(report.estimator_, RandomForestClassifier)
-    assert report.X_train is None
-    assert report.y_train is None
-    assert report.X_test is X
-    assert report.y_test is y
-
-    with pytest.raises(AttributeError):
-        report.estimator_ = LinearRegression()
-    with pytest.raises(AttributeError):
-        report.X_train = X
-    with pytest.raises(AttributeError):
-        report.y_train = y
-
-
-def test_estimator_report_from_fitted_pipeline(binary_classification_data_pipeline):
-    """Check the general behaviour of passing an already fitted pipeline without
-    refitting it.
-    """
-    estimator, X, y = binary_classification_data_pipeline
-    report = EstimatorReport(estimator, X_test=X, y_test=y)
-
-    check_is_fitted(report.estimator_)
-    assert isinstance(report.estimator_, Pipeline)
-    assert report.estimator_name_ == estimator[-1].__class__.__name__
-    assert report.X_train is None
-    assert report.y_train is None
-    assert report.X_test is X
-    assert report.y_test is y
-
-
-def test_estimator_report_invalidate_cache_data(binary_classification_data):
-    """Check that we invalidate the cache when the data is changed."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-
-    for attribute in ("X_test", "y_test"):
-        report._cache["mocking"] = "mocking"  # mock writing to cache
-        setattr(report, attribute, None)
-        assert report._cache == {}
-
-
-@pytest.mark.parametrize(
-    "Estimator, X_test, y_test, supported_plot_methods, not_supported_plot_methods",
-    [
-        (
-            RandomForestClassifier(),
-            *make_classification(random_state=42),
-            ["roc", "precision_recall"],
-            ["prediction_error"],
-        ),
-        (
-            RandomForestClassifier(),
-            *make_classification(n_classes=3, n_clusters_per_class=1, random_state=42),
-            ["roc", "precision_recall"],
-            ["prediction_error"],
-        ),
-        (
-            LinearRegression(),
-            *make_regression(random_state=42),
-            ["prediction_error"],
-            ["roc", "precision_recall"],
-        ),
-    ],
-)
-def test_estimator_report_check_support_plot(
-    Estimator, X_test, y_test, supported_plot_methods, not_supported_plot_methods
-):
-    """Check that the available plot methods are correctly registered."""
-    estimator = Estimator.fit(X_test, y_test)
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-
-    for supported_plot_method in supported_plot_methods:
-        assert hasattr(report.metrics, supported_plot_method)
-
-    for not_supported_plot_method in not_supported_plot_methods:
-        assert not hasattr(report.metrics, not_supported_plot_method)
-
-
-def test_estimator_report_help(capsys, binary_classification_data):
+def test_estimator_summarize_help(capsys, forest_binary_classification_with_test):
     """Check that the help method writes to the console."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-
-    report.help()
-    captured = capsys.readouterr()
-    assert f"Tools to diagnose estimator {estimator.__class__.__name__}" in captured.out
-
-    # Check that we have a line with accuracy and the arrow associated with it
-    assert re.search(
-        r"\.accuracy\([^)]*\).*\(↗︎\).*-.*accuracy", captured.out, re.MULTILINE
-    )
-
-
-def test_estimator_report_repr(binary_classification_data):
-    """Check that __repr__ returns a string starting with the expected prefix."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-
-    repr_str = repr(report)
-    assert "EstimatorReport" in repr_str
-
-
-@pytest.mark.parametrize(
-    "fixture_name, pass_train_data, expected_n_keys",
-    [
-        ("binary_classification_data", True, 10),
-        ("binary_classification_data_svc", True, 10),
-        ("multiclass_classification_data", True, 12),
-        ("regression_data", True, 4),
-        ("binary_classification_data", False, 5),
-        ("binary_classification_data_svc", False, 5),
-        ("multiclass_classification_data", False, 6),
-        ("regression_data", False, 2),
-    ],
-)
-@pytest.mark.parametrize("n_jobs", [1, 2])
-def test_estimator_report_cache_predictions(
-    request, fixture_name, pass_train_data, expected_n_keys, n_jobs
-):
-    """Check that calling cache_predictions fills the cache."""
-    estimator, X_test, y_test = request.getfixturevalue(fixture_name)
-    if pass_train_data:
-        report = EstimatorReport(
-            estimator, X_train=X_test, y_train=y_test, X_test=X_test, y_test=y_test
-        )
-    else:
-        report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-
-    assert report._cache == {}
-    report.cache_predictions(n_jobs=n_jobs)
-    assert len(report._cache) == expected_n_keys
-    assert report._cache != {}
-    stored_cache = deepcopy(report._cache)
-    report.cache_predictions(n_jobs=n_jobs)
-    # check that the keys are exactly the same
-    assert report._cache.keys() == stored_cache.keys()
-
-
-def test_estimator_report_pickle(binary_classification_data):
-    """Check that we can pickle an estimator report.
-
-    In particular, the progress bar from rich are pickable, therefore we trigger
-    the progress bar to be able to test that the progress bar is pickable.
-    """
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-    report.cache_predictions()
-
-    with BytesIO() as stream:
-        joblib.dump(report, stream)
-
-
-def test_estimator_report_flat_index(binary_classification_data):
-    """Check that the index is flattened when `flat_index` is True.
-
-    Since `pos_label` is None, then by default a MultiIndex would be returned.
-    Here, we force to have a single-index by passing `flat_index=True`.
-    """
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-    result = report.metrics.summarize(flat_index=True).frame()
-    assert result.shape == (8, 1)
-    assert isinstance(result.index, pd.Index)
-    assert result.index.tolist() == [
-        "precision_0",
-        "precision_1",
-        "recall_0",
-        "recall_1",
-        "roc_auc",
-        "brier_score",
-        "fit_time_s",
-        "predict_time_s",
-    ]
-
-    assert result.columns.tolist() == ["RandomForestClassifier"]
-
-
-def test_estimator_report_get_predictions():
-    """Check the behaviour of the `get_predictions` method.
-
-    We use the binary classification because it uses all the parameters of the
-    `get_predictions` method.
-    """
-    X, y = make_classification(n_classes=2, random_state=42)
-    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)
-    estimator = LogisticRegression()
-    report = EstimatorReport(
-        estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
-    )
-
-    # check the `predict` method
-    predictions = report.get_predictions(data_source="test")
-    np.testing.assert_allclose(predictions, report.estimator_.predict(X_test))
-    predictions = report.get_predictions(data_source="train")
-    np.testing.assert_allclose(predictions, report.estimator_.predict(X_train))
-    predictions = report.get_predictions(data_source="X_y", X=X_test)
-    np.testing.assert_allclose(predictions, report.estimator_.predict(X_test))
-
-    # check the validity of the `predict_proba` method
-    predictions = report.get_predictions(
-        data_source="test", response_method="predict_proba"
-    )
-    np.testing.assert_allclose(
-        predictions, report.estimator_.predict_proba(X_test)[:, 1]
-    )
-    predictions = report.get_predictions(
-        data_source="train", response_method="predict_proba", pos_label=0
-    )
-    np.testing.assert_allclose(
-        predictions, report.estimator_.predict_proba(X_train)[:, 0]
-    )
-    predictions = report.get_predictions(
-        data_source="X_y", response_method="predict_proba", X=X_test
-    )
-    np.testing.assert_allclose(
-        predictions, report.estimator_.predict_proba(X_test)[:, 1]
-    )
-
-    # check the validity of the `decision_function` method
-    predictions = report.get_predictions(
-        data_source="test", response_method="decision_function"
-    )
-    np.testing.assert_allclose(predictions, report.estimator_.decision_function(X_test))
-    predictions = report.get_predictions(
-        data_source="train", response_method="decision_function", pos_label=0
-    )
-    np.testing.assert_allclose(
-        predictions, -report.estimator_.decision_function(X_train)
-    )
-    predictions = report.get_predictions(
-        data_source="X_y", response_method="decision_function", X=X_test
-    )
-    np.testing.assert_allclose(predictions, report.estimator_.decision_function(X_test))
-
-    # check the behaviour in conjunction of a report `pos_label`
-    report = EstimatorReport(
-        estimator,
-        X_train=X_train,
-        y_train=y_train,
-        X_test=X_test,
-        y_test=y_test,
-        pos_label=0,
-    )
-    predictions = report.get_predictions(
-        data_source="train", response_method="predict_proba"
-    )
-    np.testing.assert_allclose(
-        predictions, report.estimator_.predict_proba(X_train)[:, 0]
-    )
-
-
-def test_estimator_report_get_predictions_error():
-    """Check that we raise an error when the data source is invalid."""
-    X, y = make_classification(n_classes=2, random_state=42)
-    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)
-    estimator = LogisticRegression()
-    report = EstimatorReport(
-        estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
-    )
-
-    with pytest.raises(ValueError, match="Invalid data source"):
-        report.get_predictions(data_source="invalid")
-
-
-########################################################################################
-# Check the plot methods
-########################################################################################
-
-
-def test_estimator_report_plot_roc(binary_classification_data):
-    """Check that the ROC plot method works."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-    assert isinstance(report.metrics.roc(), RocCurveDisplay)
-
-
-@pytest.mark.parametrize("display", ["roc", "precision_recall"])
-def test_estimator_report_display_binary_classification(
-    pyplot, binary_classification_data, display
-):
-    """The call to display functions should be cached."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-    assert hasattr(report.metrics, display)
-    display_first_call = getattr(report.metrics, display)()
-    assert report._cache != {}
-    display_second_call = getattr(report.metrics, display)()
-    assert display_first_call is display_second_call
-
-
-@pytest.mark.parametrize("metric", ["roc", "precision_recall"])
-def test_estimator_report_display_binary_classification_pos_label(
-    pyplot, binary_classification_data, metric
-):
-    """Check the behaviour of the display methods when `pos_label` needs to be set."""
-    X, y = make_classification(
-        n_classes=2, class_sep=0.8, weights=[0.4, 0.6], random_state=0
-    )
-    labels = np.array(["A", "B"], dtype=object)
-    y = labels[y]
-    classifier = LogisticRegression().fit(X, y)
-    report = EstimatorReport(classifier, X_test=X, y_test=y)
-    with pytest.raises(ValueError, match="pos_label is not specified"):
-        getattr(report.metrics, metric)()
-
-    report = EstimatorReport(classifier, X_test=X, y_test=y, pos_label="A")
-    display = getattr(report.metrics, metric)()
-    display.plot()
-    assert "Positive label: A" in display.ax_.get_xlabel()
-
-    display = getattr(report.metrics, metric)(pos_label="B")
-    display.plot()
-    assert "Positive label: B" in display.ax_.get_xlabel()
-
-
-@pytest.mark.parametrize("display", ["prediction_error"])
-def test_estimator_report_display_regression(pyplot, regression_data, display):
-    """The call to display functions should be cached, as long as the arguments make it
-    reproducible."""
-    estimator, X_test, y_test = regression_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-    assert hasattr(report.metrics, display)
-    display_first_call = getattr(report.metrics, display)(seed=0)
-    assert report._cache != {}
-    display_second_call = getattr(report.metrics, display)(seed=0)
-    assert display_first_call is display_second_call
-
-
-@pytest.mark.parametrize("display", ["roc", "precision_recall"])
-def test_estimator_report_display_binary_classification_external_data(
-    pyplot, binary_classification_data, display
-):
-    """The call to display functions should be cached when passing external data."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator)
-    assert hasattr(report.metrics, display)
-    display_first_call = getattr(report.metrics, display)(
-        data_source="X_y", X=X_test, y=y_test
-    )
-    assert report._cache != {}
-    display_second_call = getattr(report.metrics, display)(
-        data_source="X_y", X=X_test, y=y_test
-    )
-    assert display_first_call is display_second_call
-
-
-@pytest.mark.parametrize("display", ["prediction_error"])
-def test_estimator_report_display_regression_external_data(
-    pyplot, regression_data, display
-):
-    """The call to display functions should be cached when passing external data,
-    as long as the arguments make it reproducible."""
-    estimator, X_test, y_test = regression_data
-    report = EstimatorReport(estimator)
-    assert hasattr(report.metrics, display)
-    display_first_call = getattr(report.metrics, display)(
-        data_source="X_y", X=X_test, y=y_test, seed=0
-    )
-    assert report._cache != {}
-    display_second_call = getattr(report.metrics, display)(
-        data_source="X_y", X=X_test, y=y_test, seed=0
-    )
-    assert display_first_call is display_second_call
-
-
-@pytest.mark.parametrize("display", ["roc", "precision_recall"])
-def test_estimator_report_display_binary_classification_switching_data_source(
-    pyplot, binary_classification_data, display
-):
-    """Check that we don't hit the cache when switching the data source."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(
-        estimator, X_train=X_test, y_train=y_test, X_test=X_test, y_test=y_test
-    )
-    assert hasattr(report.metrics, display)
-    display_first_call = getattr(report.metrics, display)(data_source="test")
-    assert report._cache != {}
-    display_second_call = getattr(report.metrics, display)(data_source="train")
-    assert display_first_call is not display_second_call
-    display_third_call = getattr(report.metrics, display)(
-        data_source="X_y", X=X_test, y=y_test
-    )
-    assert display_first_call is not display_third_call
-    assert display_second_call is not display_third_call
-
-
-@pytest.mark.parametrize("display", ["prediction_error"])
-def test_estimator_report_display_regression_switching_data_source(
-    pyplot, regression_data, display
-):
-    """Check that we don't hit the cache when switching the data source."""
-    estimator, X_test, y_test = regression_data
-    report = EstimatorReport(
-        estimator, X_train=X_test, y_train=y_test, X_test=X_test, y_test=y_test
-    )
-    assert hasattr(report.metrics, display)
-    display_first_call = getattr(report.metrics, display)(data_source="test")
-    assert report._cache != {}
-    display_second_call = getattr(report.metrics, display)(data_source="train")
-    assert display_first_call is not display_second_call
-    display_third_call = getattr(report.metrics, display)(
-        data_source="X_y", X=X_test, y=y_test
-    )
-    assert display_first_call is not display_third_call
-    assert display_second_call is not display_third_call
-
-
-########################################################################################
-# Check the metrics methods
-########################################################################################
-
-
-def test_estimator_summarize_help(capsys, binary_classification_data):
-    """Check that the help method writes to the console."""
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     report.metrics.help()
@@ -604,9 +35,9 @@ def test_estimator_summarize_help(capsys, binary_classification_data):
     assert "Available metrics methods" in captured.out
 
 
-def test_estimator_summarize_repr(binary_classification_data):
+def test_estimator_summarize_repr(forest_binary_classification_with_test):
     """Check that __repr__ returns a string starting with the expected prefix."""
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     repr_str = repr(report.metrics)
@@ -615,11 +46,13 @@ def test_estimator_summarize_repr(binary_classification_data):
 
 
 @pytest.mark.parametrize("metric", ["accuracy", "brier_score", "roc_auc", "log_loss"])
-def test_estimator_summarize_binary_classification(binary_classification_data, metric):
+def test_estimator_summarize_binary_classification(
+    forest_binary_classification_with_test, metric
+):
     """Check the behaviour of the metrics methods available for binary
     classification.
     """
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     assert hasattr(report.metrics, metric)
     result = getattr(report.metrics, metric)()
@@ -644,12 +77,12 @@ def test_estimator_summarize_binary_classification(binary_classification_data, m
 
 @pytest.mark.parametrize("metric", ["precision", "recall"])
 def test_estimator_summarize_binary_classification_pr(
-    binary_classification_data, metric
+    forest_binary_classification_with_test, metric
 ):
     """Check the behaviour of the precision and recall metrics available for binary
     classification.
     """
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     assert hasattr(report.metrics, metric)
     result = getattr(report.metrics, metric)()
@@ -673,9 +106,9 @@ def test_estimator_summarize_binary_classification_pr(
 
 
 @pytest.mark.parametrize("metric", ["r2", "rmse"])
-def test_estimator_summarize_regression(regression_data, metric):
+def test_estimator_summarize_regression(linear_regression_with_test, metric):
     """Check the behaviour of the metrics methods available for regression."""
-    estimator, X_test, y_test = regression_data
+    estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     assert hasattr(report.metrics, metric)
     result = getattr(report.metrics, metric)()
@@ -729,8 +162,8 @@ def _check_results_summarize(result, expected_metrics, expected_nb_stats):
 @pytest.mark.parametrize("pos_label, nb_stats", [(None, 2), (1, 1)])
 @pytest.mark.parametrize("data_source", ["test", "X_y"])
 def test_estimator_report_summarize_binary(
-    binary_classification_data,
-    binary_classification_data_svc,
+    forest_binary_classification_with_test,
+    svc_binary_classification_with_test,
     pos_label,
     nb_stats,
     data_source,
@@ -739,7 +172,7 @@ def test_estimator_report_summarize_binary(
     classification. We test both with an SVC that does not support `predict_proba` and a
     RandomForestClassifier that does.
     """
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     kwargs = {"X": X_test, "y": y_test} if data_source == "X_y" else {}
     result = report.metrics.summarize(
@@ -783,7 +216,7 @@ def test_estimator_report_summarize_binary(
     expected_nb_stats = 2 * nb_stats + 4
     _check_results_summarize(result, expected_metrics, expected_nb_stats)
 
-    estimator, X_test, y_test = binary_classification_data_svc
+    estimator, X_test, y_test = svc_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     kwargs = {"X": X_test, "y": y_test} if data_source == "X_y" else {}
     result = report.metrics.summarize(
@@ -798,12 +231,14 @@ def test_estimator_report_summarize_binary(
 
 @pytest.mark.parametrize("data_source", ["test", "X_y"])
 def test_estimator_report_summarize_multiclass(
-    multiclass_classification_data, multiclass_classification_data_svc, data_source
+    forest_multiclass_classification_with_test,
+    svc_multiclass_classification_with_test,
+    data_source,
 ):
     """Check the behaviour of the `summarize` method with multiclass
     classification.
     """
-    estimator, X_test, y_test = multiclass_classification_data
+    estimator, X_test, y_test = forest_multiclass_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     kwargs = {"X": X_test, "y": y_test} if data_source == "X_y" else {}
     result = report.metrics.summarize(data_source=data_source, **kwargs).frame()
@@ -821,7 +256,7 @@ def test_estimator_report_summarize_multiclass(
     expected_nb_stats = 3 * 3 + 3
     _check_results_summarize(result, expected_metrics, expected_nb_stats)
 
-    estimator, X_test, y_test = multiclass_classification_data_svc
+    estimator, X_test, y_test = svc_multiclass_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     kwargs = {"X": X_test, "y": y_test} if data_source == "X_y" else {}
     result = report.metrics.summarize(data_source=data_source, **kwargs).frame()
@@ -833,9 +268,11 @@ def test_estimator_report_summarize_multiclass(
 
 
 @pytest.mark.parametrize("data_source", ["test", "X_y"])
-def test_estimator_report_summarize_regression(regression_data, data_source):
+def test_estimator_report_summarize_regression(
+    linear_regression_with_test, data_source
+):
     """Check the behaviour of the `summarize` method with regression."""
-    estimator, X_test, y_test = regression_data
+    estimator, X_test, y_test = linear_regression_with_test
     kwargs = {"X": X_test, "y": y_test} if data_source == "X_y" else {}
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     result = report.metrics.summarize(data_source=data_source, **kwargs).frame()
@@ -845,10 +282,10 @@ def test_estimator_report_summarize_regression(regression_data, data_source):
 
 
 def test_estimator_report_summarize_scoring_kwargs(
-    regression_multioutput_data, multiclass_classification_data
+    linear_regression_multioutput_with_test, forest_multiclass_classification_with_test
 ):
     """Check the behaviour of the `summarize` method with scoring kwargs."""
-    estimator, X_test, y_test = regression_multioutput_data
+    estimator, X_test, y_test = linear_regression_multioutput_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     assert hasattr(report.metrics, "summarize")
     result = report.metrics.summarize(
@@ -858,7 +295,7 @@ def test_estimator_report_summarize_scoring_kwargs(
     assert isinstance(result.index, pd.MultiIndex)
     assert result.index.names == ["Metric", "Output"]
 
-    estimator, X_test, y_test = multiclass_classification_data
+    estimator, X_test, y_test = forest_multiclass_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     assert hasattr(report.metrics, "summarize")
     result = report.metrics.summarize(scoring_kwargs={"average": None}).frame()
@@ -871,17 +308,17 @@ def test_estimator_report_summarize_scoring_kwargs(
     "fixture_name, scoring_names, expected_columns",
     [
         (
-            "regression_data",
+            "linear_regression_with_test",
             ["R2", "RMSE", "FIT_TIME", "PREDICT_TIME"],
             ["R2", "RMSE", "FIT_TIME", "PREDICT_TIME"],
         ),
         (
-            "regression_data",
+            "linear_regression_with_test",
             ["R2", "RMSE", "FIT_TIME", "PREDICT_TIME"],
             ["R2", "RMSE", "FIT_TIME", "PREDICT_TIME"],
         ),
         (
-            "multiclass_classification_data",
+            "forest_multiclass_classification_with_test",
             ["Precision", "Recall", "ROC AUC", "Log Loss", "Fit Time", "Predict Time"],
             [
                 "Precision",
@@ -919,10 +356,10 @@ def test_estimator_report_summarize_overwrite_scoring_names(
 
 
 def test_estimator_report_summarize_indicator_favorability(
-    binary_classification_data,
+    forest_binary_classification_with_test,
 ):
     """Check that the behaviour of `indicator_favorability` is correct."""
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     result = report.metrics.summarize(indicator_favorability=True).frame()
     assert "Favorability" in result.columns
@@ -943,11 +380,11 @@ def test_estimator_report_summarize_indicator_favorability(
     ],
 )
 def test_estimator_report_summarize_scoring_single_list_equivalence(
-    binary_classification_data, scoring, scoring_names, scoring_kwargs
+    forest_binary_classification_with_test, scoring, scoring_names, scoring_kwargs
 ):
     """Check that passing a single string, callable, scorer is equivalent to passing a
     list with a single element, and it's possible to overwrite col name."""
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     result_single = report.metrics.summarize(
         scoring=scoring, scoring_names=scoring_names, scoring_kwargs=scoring_kwargs
@@ -959,9 +396,11 @@ def test_estimator_report_summarize_scoring_single_list_equivalence(
     assert result_single.equals(result_list)
 
 
-def test_estimator_report_interaction_cache_metrics(regression_multioutput_data):
+def test_estimator_report_interaction_cache_metrics(
+    linear_regression_multioutput_with_test,
+):
     """Check that the cache take into account the 'kwargs' of a metric."""
-    estimator, X_test, y_test = regression_multioutput_data
+    estimator, X_test, y_test = linear_regression_multioutput_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     # The underlying metrics will call `_compute_metric_scores` that take some arbitrary
@@ -992,9 +431,9 @@ def test_estimator_report_interaction_cache_metrics(regression_multioutput_data)
     assert isinstance(result_r2_uniform_average, float)
 
 
-def test_estimator_report_custom_metric(regression_data):
+def test_estimator_report_custom_metric(linear_regression_with_test):
     """Check the behaviour of the `custom_metric` computation in the report."""
-    estimator, X_test, y_test = regression_data
+    estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     def custom_metric(y_true, y_pred, threshold=0.5):
@@ -1043,20 +482,24 @@ def test_estimator_report_custom_metric(regression_data):
 
 
 @pytest.mark.parametrize("scoring", ["public_metric", "_private_metric"])
-def test_estimator_report_summarize_error_scoring_strings(regression_data, scoring):
+def test_estimator_report_summarize_error_scoring_strings(
+    linear_regression_with_test, scoring
+):
     """Check that we raise an error if a scoring string is not a valid metric."""
-    estimator, X_test, y_test = regression_data
+    estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     err_msg = re.escape(f"Invalid metric: {scoring!r}.")
     with pytest.raises(ValueError, match=err_msg):
         report.metrics.summarize(scoring=[scoring])
 
 
-def test_estimator_report_custom_function_kwargs_numpy_array(regression_data):
+def test_estimator_report_custom_function_kwargs_numpy_array(
+    linear_regression_with_test,
+):
     """Check that we are able to store a hash of a numpy array in the cache when they
     are passed as kwargs.
     """
-    estimator, X_test, y_test = regression_data
+    estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     weights = np.ones_like(y_test) * 2
     hash_weights = joblib.hash(weights)
@@ -1084,10 +527,10 @@ def test_estimator_report_custom_function_kwargs_numpy_array(regression_data):
     )
 
 
-def test_estimator_report_summarize_with_custom_metric(regression_data):
+def test_estimator_report_summarize_with_custom_metric(linear_regression_with_test):
     """Check that we can pass a custom metric with specific kwargs into
     `summarize`."""
-    estimator, X_test, y_test = regression_data
+    estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     weights = np.ones_like(y_test) * 2
 
@@ -1108,10 +551,10 @@ def test_estimator_report_summarize_with_custom_metric(regression_data):
     )
 
 
-def test_estimator_report_summarize_with_scorer(regression_data):
+def test_estimator_report_summarize_with_scorer(linear_regression_with_test):
     """Check that we can pass scikit-learn scorer with different parameters to
     the `summarize` method."""
-    estimator, X_test, y_test = regression_data
+    estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     weights = np.ones_like(y_test) * 2
 
@@ -1140,14 +583,14 @@ def test_estimator_report_summarize_with_scorer(regression_data):
 
 
 def test_estimator_report_custom_metric_compatible_estimator(
-    binary_classification_data,
+    forest_binary_classification_with_test,
 ):
     """Check that the estimator report still works if an estimator has a compatible
     scikit-learn API.
     """
-    _, X_test, y_test = binary_classification_data
+    _, X_test, y_test = forest_binary_classification_with_test
 
-    class CompatibleEstimator:
+    class CompatibleEstimator(BaseEstimator):
         """Estimator exposing only a predict method but it should be enough for the
         reports.
         """
@@ -1188,7 +631,7 @@ def test_estimator_report_custom_metric_compatible_estimator(
     ],
 )
 def test_estimator_report_summarize_with_scorer_binary_classification(
-    binary_classification_data, scorer, pos_label
+    forest_binary_classification_with_test, scorer, pos_label
 ):
     """Check that we can pass scikit-learn scorer with different parameters to
     the `summarize` method.
@@ -1196,7 +639,7 @@ def test_estimator_report_summarize_with_scorer_binary_classification(
     We also check that we can pass `pos_label` whether to the scorer or to the
     `summarize` method or consistently to both.
     """
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     result = report.metrics.summarize(
@@ -1209,17 +652,24 @@ def test_estimator_report_summarize_with_scorer_binary_classification(
         [
             [accuracy_score(y_test, estimator.predict(X_test))],
             [accuracy_score(y_test, estimator.predict(X_test))],
-            [f1_score(y_test, estimator.predict(X_test), average="macro", pos_label=1)],
+            [
+                f1_score(
+                    y_test,
+                    estimator.predict(X_test),
+                    average="macro",
+                    pos_label=pos_label,
+                )
+            ],
         ],
     )
 
 
 def test_estimator_report_summarize_with_scorer_pos_label_error(
-    binary_classification_data,
+    forest_binary_classification_with_test,
 ):
     """Check that we raise an error when pos_label is passed both in the scorer and
     globally conducting to a mismatch."""
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     f1_scorer = make_scorer(
@@ -1232,9 +682,9 @@ def test_estimator_report_summarize_with_scorer_pos_label_error(
         report.metrics.summarize(scoring=[f1_scorer], pos_label=0).frame()
 
 
-def test_estimator_report_summarize_invalid_metric_type(regression_data):
+def test_estimator_report_summarize_invalid_metric_type(linear_regression_with_test):
     """Check that we raise the expected error message if an invalid metric is passed."""
-    estimator, X_test, y_test = regression_data
+    estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     err_msg = re.escape("Invalid type of metric: <class 'int'> for 1")
@@ -1416,13 +866,13 @@ def test_estimator_report_brier_score_requires_binary_classification():
     assert not hasattr(report.metrics, "brier_score")
 
 
-def test_estimator_report_average_return_float(binary_classification_data):
+def test_estimator_report_average_return_float(forest_binary_classification_with_test):
     """Check that we expect a float value when computing a metric with averaging.
 
     Non-regression test for:
     https://github.com/probabl-ai/skore/issues/1501
     """
-    estimator, X_test, y_test = binary_classification_data
+    estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     for metric_name in ("precision", "recall", "roc_auc"):
@@ -1430,9 +880,11 @@ def test_estimator_report_average_return_float(binary_classification_data):
         assert isinstance(result, float)
 
 
-def test_estimator_report_metric_with_neg_metrics(binary_classification_data):
+def test_estimator_report_metric_with_neg_metrics(
+    forest_binary_classification_with_test,
+):
     """Check that scikit-learn metrics with 'neg_' prefix are handled correctly."""
-    classifier, X_test, y_test = binary_classification_data
+    classifier, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
 
     result = report.metrics.summarize(scoring=["neg_log_loss"]).frame()
@@ -1442,9 +894,11 @@ def test_estimator_report_metric_with_neg_metrics(binary_classification_data):
     )
 
 
-def test_estimator_report_with_sklearn_scoring_strings(binary_classification_data):
+def test_estimator_report_with_sklearn_scoring_strings(
+    forest_binary_classification_with_test,
+):
     """Test that scikit-learn metric strings can be passed to summarize."""
-    classifier, X_test, y_test = binary_classification_data
+    classifier, X_test, y_test = forest_binary_classification_with_test
     class_report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
 
     result = class_report.metrics.summarize(scoring=["neg_log_loss"]).frame()
@@ -1463,9 +917,11 @@ def test_estimator_report_with_sklearn_scoring_strings(binary_classification_dat
     assert favorability == "(↘︎)"
 
 
-def test_estimator_report_with_sklearn_scoring_strings_regression(regression_data):
+def test_estimator_report_with_sklearn_scoring_strings_regression(
+    linear_regression_with_test,
+):
     """Test scikit-learn regression metric strings in summarize."""
-    regressor, X_test, y_test = regression_data
+    regressor, X_test, y_test = linear_regression_with_test
     reg_report = EstimatorReport(regressor, X_test=X_test, y_test=y_test)
 
     reg_result = reg_report.metrics.summarize(
@@ -1481,9 +937,9 @@ def test_estimator_report_with_sklearn_scoring_strings_regression(regression_dat
     assert reg_result.loc["R²"]["Favorability"] == "(↗︎)"
 
 
-def test_estimator_report_with_scoring_strings_regression(regression_data):
+def test_estimator_report_with_scoring_strings_regression(linear_regression_with_test):
     """Test scikit-learn regression metric strings in summarize."""
-    regressor, X_test, y_test = regression_data
+    regressor, X_test, y_test = linear_regression_with_test
     reg_report = EstimatorReport(regressor, X_test=X_test, y_test=y_test)
 
     reg_result = reg_report.metrics.summarize(
@@ -1499,9 +955,11 @@ def test_estimator_report_with_scoring_strings_regression(regression_data):
     assert reg_result.loc["R²"]["Favorability"] == "(↗︎)"
 
 
-def test_estimator_report_sklearn_scorer_names_pos_label(binary_classification_data):
+def test_estimator_report_sklearn_scorer_names_pos_label(
+    forest_binary_classification_with_test,
+):
     """Check that `pos_label` is dispatched with scikit-learn scorer names."""
-    classifier, X_test, y_test = binary_classification_data
+    classifier, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
 
     result = report.metrics.summarize(scoring=["f1"], pos_label=0).frame()
@@ -1516,12 +974,12 @@ def test_estimator_report_sklearn_scorer_names_pos_label(binary_classification_d
 
 
 def test_estimator_report_sklearn_scorer_names_scoring_kwargs(
-    binary_classification_data,
+    forest_binary_classification_with_test,
 ):
     """Check that `scoring_kwargs` is not supported when `scoring` is a scikit-learn
     scorer name.
     """
-    classifier, X_test, y_test = binary_classification_data
+    classifier, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
 
     err_msg = (
@@ -1537,9 +995,7 @@ def test_estimator_report_sklearn_scorer_names_scoring_kwargs(
 @pytest.mark.parametrize(
     "metric, metric_fn", [("precision", precision_score), ("recall", recall_score)]
 )
-def test_estimator_report_summarize_pos_label_overwrite(
-    binary_classification_data, metric, metric_fn
-):
+def test_estimator_report_summarize_pos_label_overwrite(metric, metric_fn):
     """Check that `pos_label` can be overwritten in `summarize`"""
     X, y = make_classification(
         n_classes=2, class_sep=0.8, weights=[0.4, 0.6], random_state=0
@@ -1571,9 +1027,7 @@ def test_estimator_report_summarize_pos_label_overwrite(
 @pytest.mark.parametrize(
     "metric, metric_fn", [("precision", precision_score), ("recall", recall_score)]
 )
-def test_estimator_report_precision_recall_pos_label_overwrite(
-    binary_classification_data, metric, metric_fn
-):
+def test_estimator_report_precision_recall_pos_label_overwrite(metric, metric_fn):
     """Check that `pos_label` can be overwritten in `summarize`"""
     X, y = make_classification(
         n_classes=2, class_sep=0.8, weights=[0.4, 0.6], random_state=0
@@ -1596,7 +1050,7 @@ def test_estimator_report_precision_recall_pos_label_overwrite(
 
 
 def test_estimator_report_roc_multiclass_requires_predict_proba(
-    multiclass_classification_data_svc, binary_classification_data_svc
+    svc_multiclass_classification_with_test, svc_binary_classification_with_test
 ):
     """Check that the ROC AUC metric is not exposed with multiclass data if the
     estimator does not expose `predict_proba`.
@@ -1604,190 +1058,12 @@ def test_estimator_report_roc_multiclass_requires_predict_proba(
     Non-regression test for:
     https://github.com/probabl-ai/skore/issues/1873
     """
-    classifier, X_test, y_test = multiclass_classification_data_svc
+    classifier, X_test, y_test = svc_multiclass_classification_with_test
     report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
     with pytest.raises(AttributeError):
         report.metrics.roc_auc()
 
-    classifier, X_test, y_test = binary_classification_data_svc
+    classifier, X_test, y_test = svc_binary_classification_with_test
     report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
     assert hasattr(report.metrics, "roc_auc")
     report.metrics.roc_auc()
-
-
-########################################################################################
-# Check the data accessor methods
-########################################################################################
-
-
-@pytest.mark.parametrize(
-    "params, err_msg",
-    [
-        (dict(data_source="invalid"), "'data_source' options are"),
-        (dict(subsample_strategy="invalid"), "'subsample_strategy' options are"),
-    ],
-)
-def test_estimator_report_analyze_error(binary_classification_data, params, err_msg):
-    """Check that the `analyze` method raises an error when the data source is not
-    valid."""
-    classifier, X_test, y_test = binary_classification_data
-    report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
-    with pytest.raises(ValueError, match=err_msg):
-        report.data.analyze(**params)
-
-
-def test_estimator_report_analyze_data_source_not_available(binary_classification_data):
-    """Check that we raise a proper error message when the data source requested is
-    not available."""
-    classifier, X_test, y_test = binary_classification_data
-    report = EstimatorReport(classifier)
-
-    err_msg = "X_train is required when `data_source=train`"
-    with pytest.raises(ValueError, match=err_msg):
-        report.data.analyze(data_source="train")
-
-    err_msg = "X_test is required when `data_source=test`"
-    with pytest.raises(ValueError, match=err_msg):
-        report.data.analyze(data_source="test")
-
-    err_msg = "X_train is required when `data_source=all`"
-    with pytest.raises(ValueError, match=err_msg):
-        report.data.analyze(data_source="all")
-
-    # if not requesting `y`, we should not raise an error
-    report = EstimatorReport(classifier, X_test=X_test)
-    display = report.data.analyze(data_source="test", with_y=False)
-    np.testing.assert_array_equal(display.summary["dataframe"].to_numpy(), X_test)
-
-    err_msg = "y_test is required when `data_source=test`"
-    with pytest.raises(ValueError, match=err_msg):
-        report.data.analyze(data_source="test", with_y=True)
-
-
-def test_estimator_report_analyze_data_source_without_y():
-    """Check the behaviour of `data_source` parameter without including `y`."""
-    X, y = make_classification(n_classes=2, random_state=42)
-    X = pd.DataFrame(X, columns=[f"Column {i}" for i in range(X.shape[1])])
-    y = pd.Series(y, name="Classification target")
-    X_train, X_test = X.iloc[:100], X.iloc[100:]
-    y_train, y_test = y.iloc[:100], y.iloc[100:]
-    classifier = LogisticRegression()
-
-    report = EstimatorReport(
-        classifier, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
-    )
-
-    display = report.data.analyze(data_source="train", with_y=False)
-    pd.testing.assert_frame_equal(display.summary["dataframe"], X_train)
-
-    display = report.data.analyze(data_source="test", with_y=False)
-    pd.testing.assert_frame_equal(display.summary["dataframe"], X_test)
-
-    display = report.data.analyze(data_source="all", with_y=False)
-    pd.testing.assert_frame_equal(display.summary["dataframe"], X)
-
-
-def test_estimator_report_analyze_data_source_with_y():
-    """Check the behaviour of `data_source` parameter with `y`."""
-    X, y = make_classification(n_classes=2, random_state=42)
-    X = pd.DataFrame(X, columns=[f"Column {i}" for i in range(X.shape[1])])
-    y = pd.Series(y, name="Classification target")
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, shuffle=False, train_size=50
-    )
-    classifier = LogisticRegression()
-
-    report = EstimatorReport(
-        classifier, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
-    )
-
-    display = report.data.analyze(data_source="train")
-    pd.testing.assert_frame_equal(
-        display.summary["dataframe"], pd.concat([X_train, y_train], axis=1)
-    )
-
-    display = report.data.analyze(data_source="test")
-    pd.testing.assert_frame_equal(
-        display.summary["dataframe"], pd.concat([X_test, y_test], axis=1)
-    )
-
-    display = report.data.analyze(data_source="all")
-    pd.testing.assert_frame_equal(
-        display.summary["dataframe"], pd.concat([X, y], axis=1)
-    )
-
-
-@pytest.mark.parametrize("data_source", ["train", "test", "all"])
-@pytest.mark.parametrize(
-    "n_targets, target_column_names", [(1, ["Target"]), (2, ["Target 0", "Target 1"])]
-)
-def test_estimator_report_analyze_numpy_array(
-    data_source, n_targets, target_column_names
-):
-    """Check that NumPy arrays are converted to pandas DataFrames when data are
-    retrieved."""
-    X, y = make_regression(
-        n_samples=100, n_features=2, n_targets=n_targets, random_state=42
-    )
-    X_train, X_test, y_train, y_test = train_test_split(
-        X, y, shuffle=False, train_size=50
-    )
-    classifier = LinearRegression()
-
-    report = EstimatorReport(
-        classifier, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
-    )
-
-    display = report.data.analyze(data_source=data_source, with_y=False)
-    assert display.summary["dataframe"].columns.to_list() == [
-        f"Feature {i}" for i in range(X_train.shape[1])
-    ]
-
-    display = report.data.analyze(data_source=data_source, with_y=True)
-    assert (
-        display.summary["dataframe"].columns.to_list()
-        == [f"Feature {i}" for i in range(X_train.shape[1])] + target_column_names
-    )
-
-
-@pytest.mark.parametrize("subsample_strategy", ["head", "random"])
-def test_estimator_report_analyze_subsampling(
-    binary_classification_data, subsample_strategy
-):
-    """Check that the `subsample` parameter is handled correctly."""
-    classifier, X_test, y_test = binary_classification_data
-    X_test = pd.DataFrame(
-        X_test, columns=[f"Feature {i}" for i in range(X_test.shape[1])]
-    )
-    y_test = pd.Series(y_test, name="Target")
-    report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
-
-    display = report.data.analyze(
-        data_source="test", subsample=10, subsample_strategy=subsample_strategy, seed=42
-    )
-    assert display.summary["dataframe"].shape[0] == 10
-
-    if subsample_strategy == "head":
-        assert display.summary["dataframe"].index.to_list() == list(range(10))
-    else:
-        assert display.summary["dataframe"].index.to_list() != list(range(10))
-
-
-def test_estimator_data_accessor_help(capsys, binary_classification_data):
-    """Check that the help method writes to the console."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-
-    report.data.help()
-    captured = capsys.readouterr()
-    assert "Available data methods" in captured.out
-
-
-def test_estimator_data_accessor_repr(binary_classification_data):
-    """Check that __repr__ returns a string starting with the expected prefix."""
-    estimator, X_test, y_test = binary_classification_data
-    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-
-    repr_str = repr(report.data)
-    assert "skore.EstimatorReport.data" in repr_str
-    assert "help()" in repr_str
